+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
  * removed `--force` option from `puli build`
  * added `--force` option to `puli bind`
  * added `--force` option to `puli type define`
+ * renamed `puli config --delete|-d` to `puli config --reset|-r`
 
 * 1.0.0-beta3 (2015-03-19)